templates_builder/requirements.txt
beautifulsoup4~=4.12.2
soupsieve~=2.4.1
pip~=21.3.1
wheel~=0.37.1
google~=3.0.0
keyring~=23.13.1
rsa~=4.9
pyasn1~=0.5.0
rich~=13.3.5
Pygments~=2.15.1
MarkupSafe~=2.1.2
Werkzeug~=2.2.3
Jinja2~=3.1.2
click~=8.1.3
blinker~=1.6.2
itsdangerous~=2.1.2
mdurl~=0.1.2
twine~=4.0.2
requests~=2.28.2
rfc3986~=2.0.0
pkginfo~=1.9.6
urllib3~=1.26.15
six~=1.16.0
webencodings~=0.5.1
bleach~=6.0.0
cachetools~=5.3.0
certifi~=2022.12.7
idna~=3.4
docutils~=0.19
watchdog~=3.0.0
setuptools~=60.2.0
cloudevents~=1.9.0
deprecation~=2.1.0
packaging~=23.1
zipp~=3.15.0
Flask~=2.2.3
google-cloud-storage~=2.8.0
google-cloud-firestore~=2.11.1
google-cloud-pubsub~=2.16.0
google-auth~=2.17.3
functions-framework==3.3.0
PyYAML~=6.0
pluggy~=1.0.0